[HDRP] Water samples
The goal of this PR is to deliver simple, lightweight and easy to understand scenes to users so that they can copy the setup, behavior, scripts and use it as a headstart to make their own content. 

Currently the sample weighs roughly 12 mb, so this shouldn't add too much pressure on the HDRP Package weight.
![image](https://media.github.cds.internal.unity3d.com/user/1764/files/f0eb1797-10b1-47ac-9930-9ccc79f145f6)

The sample includes 4 scenes, a Pool Scene, a Glacier Scene, an Island Scene and a Water line scene.
All the scene include an "infinite" ground (or water) to avoid having a sample that is floating into space as per the new guideline for samples (rest of the sample will be reworked soon to match this)

**Pool**
![image](https://media.github.cds.internal.unity3d.com/user/1764/files/db7a2577-d9cf-4d26-8a5b-cd3112659041)

**River**
![River_20230130_185036_River](https://media.github.cds.internal.unity3d.com/user/1764/files/ed0b2a0c-1679-4ae2-a680-26b7244ddebc)

We are aware that the River sample uses an Ocean and not a River Surface Type. The idea was to showcase how to make a river flow into the Ocean for GDC. We wanted to keep the same content since a video will surely be made as a tutorial on how to make this scene so it made sense to keep Ocean even if we focus on the River here. So it made sense to give users the same content that there is on the video. 
Changing the type from Ocean to River works mostly the same besides a few tweaks. 

**Ocean**
![Ocean_20230130_1850256_Ocean2](https://media.github.cds.internal.unity3d.com/user/1764/files/7471b176-0cd6-4310-9e7a-e55e7f3565d5)

**Waterline**
![image](https://media.github.cds.internal.unity3d.com/user/1764/files/cba931e6-6899-4e66-ab55-3091ce7487e9)


Right after importing the sample and opening a scene, a popup will appear to make sure your HDRP asset is compatible with the sample. (Courtesy of @remym). That way, we don't have to override or change the HDRP Asset without users knowing. The scripts can benefits from some polish but this can be done later.  
![image](https://media.github.cds.internal.unity3d.com/user/1764/files/d7b766f4-7aa5-46a7-b564-600a6802dfb9)

Also, in each scene, there's a small description of the sample and areas in which the scene focuses. You can see the description by clicking on the "Water Sample Description" Gameobject and look at the inspector. 
![image](https://media.github.cds.internal.unity3d.com/user/1764/files/0c5a68aa-7cc1-4951-ab6b-ffa80b7d6b38)