[HDRP][Water] API to sync simulation time
Added public API on the water surface monobehavior to set the simulation start time.
That way surfaces could be synced across network for eg. a multiplayer game.

Example (will setup the water as if simulation started 10 seconds ago):
```
water.simulationStart = DateTime.Now - TimeSpan.FromSeconds(10);
```

The variable can also be read and copied to another surface created later to sync them locally
```
water1.simulationStart = water2.simulationStart;
```

Important note: this API doesn't have any effect on the first frame where the water surface is created. It could be fixed but requires some refactoring, since it doesn't seem to be a big issue i didn't do it, but it's worth mentionning it in the doc for example.