[Backport][2023.1][VFX] Multiple small UI fixes
Original PR: https://github.cds.internal.unity3d.com/unity/unity/pull/21700
https://jira.unity3d.com/browse/UUM-25106
https://jira.unity3d.com/browse/UUM-23945
https://jira.unity3d.com/browse/UUM-24248

Fixed three very small independent issues.

### UUM-25106
When dragging a texture from the project manager to an output context in the VFX Graph the texture is correctly referenced, but an exception is raised in the console.

>InvalidCastException: Specified cast is not valid.
UnityEditor.VFX.UI.VFXView.OnDragPerform (UnityEngine.UIElements.DragPerformEvent e) (at D:/github/unity/Packages/com.unity.visualeffectgraph/Editor/GraphView/Views/VFXView.cs:2989

### UUM-23945
Settings with dropdown are cropped in the UI making them unreadable.
Expected: Dropdown are not cropped and actual value is fully readable from the node UI
Tested on: 2022.2a16 but it's older than that. Not sure till when and if it's a regression.
![screenshot](https://jira.unity3d.com/secure/attachment/667602/667602_image-2022-06-08-17-36-51-659.png)

### UUM-24248
This issue comes from a user feedback on the [forum](https://forum.unity.com/threads/modifying-exposedproperties-doesnt-trigger-endchangecheck-in-custom-editor-windows.1334666/)
It's about the `GUI.changed` state in the VXF component inspector which was not reliable.
The user even provided a fix, but the proposed fix was a workaround, I fixed the root cause of the issue.