Change the behaviour of the SRP Batcher controll script in test 9603
Fix for the issue raised in [this slack discussion](https://unity.slack.com/archives/C6Y79CZM0/p1675240947098239).
DisableBatching.cs was toggling the batching state without verifications, and not settings it back afterwards.

I replaced it with a method that takes the desired state as input, and restores the original batching state on destoy (when scene unloads for the next test to run).

[Yamato Job](https://unity-ci.cds.internal.unity3d.com/job/20750264/dependency-graph)