Fix Decal Projector Editor fields not being saved on prefabs
Fixes https://jira.unity3d.com/browse/UUM-29105

The Decal Projector Editor bypasses SerializedObject in order to do some custom logic that keeps track of pivot points and apply different changes consistently when multi-editing. As a result, the code also must handle applying the edited values carefully. In this case, the prefab asset did not end up dirty and therefore it was not serialized to disk. The added code ensures the changes end up in the .prefab file. It appears the issue was introduced by this PR: https://github.com/Unity-Technologies/Graphics/pull/3600