[2023.2] Reduce import times for VFX Graph assets
This PR addresses three different performance problems that I have noticed while profiling VFX imports in a user a project. In particular, I have looked at the case where many VFX graphs are imported at once. This happens frequently when e.g. checking out a new version from version control and then opening the editor, or when some shader define that affects VFX graphs changes (- I do not understand the specific, but a client assured me that this is happening to them).

The first problem is that there is an asset postprocessor for VFX graphs. This asset postprocessor occasionally runs (not always) and takes a long time because it touches many assets and causes an asset database refresh for each single one. I have wrapped the code path with `StartAssetEditing`/`StopAssetEditing`, which will only trigger a single refresh when we are done. This has a massive impact, e.g. below you can see that instead of 37s we are now only spending 8s on this asset postprocessor.

Before:
![vfx-graph-step-0](https://media.github.cds.internal.unity3d.com/user/1745/files/39430771-a49d-4a37-9e20-8a095c47938e)

After:
![vfx-graph-on-will-save-assets-after](https://media.github.cds.internal.unity3d.com/user/1745/files/79a8a6c0-484e-4154-9177-79f568161f52)

The second problem is something that I had looked at previously already. It turns out that `StringBuilder.Replace` is ridiculously slow for reasons beyond my understanding. Speeding that up by itself is hard because it's all part of the standard library shipping with Mono. However, I was able to massage the code path a little bit more so that the function can be called with more precise bounds on where to search for the string to replace. This requires some more computation on strings to happen, but overall it's still a huge win: This function is the biggest problem in VFX import and it gets about 3x faster.

Before:
![vfx-graph-replace-before](https://media.github.cds.internal.unity3d.com/user/1745/files/dbdc7a3d-590c-4d1e-bbc5-efeaf22011d5)

After:
![vfx-graph-replace-after](https://media.github.cds.internal.unity3d.com/user/1745/files/bb16697d-99b4-4f9d-9b37-d8da88c2c102)

The last problem is not as big as the first two :) - `FindConcreteSubclasses` was manually going through all the different assemblies and was looking for types that are either a subclass of a concrete type or are marked with a specific attribute. This is quite slow. Luckily, we already have all of this information anyway in the TypeCache! So I rewrote the method in a way that makes use of the TypeCache.