Graphics/gpudriven initial analytic derivatives
The purpose of this PR is to add support for analytic partial derivatives, particularly for the GPU Driven Deferred Materials initiative. The deferred materials initiative involves rendering a triangle visibility buffer in a single pass, and rendering materials in a later pass (hence the name, deferred materials). This approach has significant performance gains for objects with many, small triangles.

One of the problems with this approach is that across triangle boundaries, the derivatives used in a Texture2D.Sample() calculation are invalid. In practice, this means that edges tend to sample the smallest (one pixel) mipmap which causes unacceptable artifacts.
![image_on](https://media.github.cds.internal.unity3d.com/user/4640/files/bb549bbd-c6e8-463b-ba96-d6d2721a5084)
![image_off](https://media.github.cds.internal.unity3d.com/user/4640/files/030268c1-153a-48b7-a447-bde0d14d9c12)

Thus, this PR analyzes the text from ShaderGraph, parses it, calculates derivatives, and passes in new text to the ShaderGraph splice dictionary.

The GPU Driven initiative is a pretty big task, so to minimize the friction we are trying to break it into smaller pieces that we can test and validate in trunk instead of having one giant PR later that is difficult to understand/test/etc.

This PR adds an option to shader graph for analytic derivatves with three options:
1. Disabled. The default option is do nothing, thus having no effect on existing projects.
2. VBuffer Only. This path will be used by Deferred Materials to enable analytic derivatives.
3. Emulate All. This is a debug option for developer testing. It creates the initial derivatives of the UVs using ddx/ddy but otherwise calculates the values analytically. This value works in both HDRP and URP, and allows us to test with real data inside trunk.

Once GPU Driven ships, the "disabled" option would be removed and the only options would be VBuffer and Emulate All. If there is a cleaner way to hide these functions (and make them clear that they are dev-only options) then I'm happy to change it.

Additionally, there is an option to dump the parsed information to a text file. The intention is that we can test shader graphs and if parsing/generation fails, we can debug and fix the issue using these text files instead of a full build of their game.

Finally, it should be noted that this PR is NOT a generic HLSL parser. There are many corner cases of HLSL that this parser does not handle, which is acceptable because this parser only runs on ShaderGraph generated nodes. User functions are explicitly ignored during parsing.