[Backport][2021.3][VFX]Fixed unmatched BeginChangeCheck call breaking the GUI.changed state
Original PR: https://github.cds.internal.unity3d.com/unity/unity/pull/21700

Jira: UUM-24248
This issue comes from a user feedback on the [forum](https://forum.unity.com/threads/modifying-exposedproperties-doesnt-trigger-endchangecheck-in-custom-editor-windows.1334666/)
It's about the `GUI.changed` state in the VXF component inspector which was not reliable.
The user even provided a fix, but the proposed fix was a workaround, I fixed the root cause of the issue.