Update GTF version and ImageAssert.AreEqual method call
The new version of `com.unity.testframework.graphics` package [(PR #154)](https://github.cds.internal.unity3d.com/unity/com.unity.testframework.graphics/pull/154) is introducing a variable to the GraphicsTestCase.cs `ReferenceImagePathLog` that logs the expected path of the reference image failed.
This PR updates the gtf version and the ImageAssert.AreEqual method calls that use the GraphicsTestCase class to provide the testCases for the tests.