[2023.2] Improve shader graph import performance
This PR optimizes shader graph imports. I came to these optimizations by profiling a user project and repeatedly reimporting their shaders. In particular, there was one shader that took 23.5s to import. With these optimizations here, it's down to 5.3s (still not great, but better).


The first optimization is not actually part of this PR anymore. I noticed that you have [already fixed](https://github.cds.internal.unity3d.com/unity/unity/pull/22666) it in February, though the same problem exists in older versions of Unity and I hope to backport parts of it soon. The crucial part here is the optimization that you applied to your `GetForest(...)` method there; it gets rid of many linear searches and replaces them with constant time hashtable look-ups. This optimization already saves a few seconds, but even with this there is still more work to do :)

Second, I added a batch method `GraphData.RemoveEdges` to remove edges. This happens during shader import (why? no idea!) and every edge that you remove triggers a revalidation of the graph. It's much more sensible to remove them all and then revalidate once, which saves a lot of time because validation is expensive.

Third, I added an overload of `GraphData.GetEdges` that takes a `MaterialSlot` instead of a `SlotReference`. This looks innocent, but when you look at `NodeUtils.GetChildNodes` you will see why this makes sense. Before my new overload, this method first took all of its `MaterialSlot` instances and got their id of type `SlotReference`, and then asked the graph to get all edges for that `SlotReference`, which internally performs a linear search for the `MaterialSlot` that the `SlotReference` represents. So we're throwing away information only to then recompute it in a very expensive manner. Not doing this is one of the major parts of this optimization, because this code path runs _a lot_ when you have a big graph.

Finally, I have looked at `NodeUtils.GetEffectiveShaderStageCapability`. This code is used in `ShaderSubGraphImporter.ProcessSubGraph` to compute the shader stage capabilities for every port. The current unoptimized version starts from scratch for every slot. This is bad, because we are duplicating a ton of work this way. The stage capabilities for every output depend on the ports that this node is connected to and we have to essentially recurse down into the graph and keep track of the capabilities of the slots we see. I have turned the algorithm on its head: Instead of just working downwards and computing the shader stage for the current slot, we work upwards and compute the shader stage for every slot that we have seen and cache it. This by itself saves a huge chunk of time, before:
![shader-graph-get-effective-shader-stage-before](https://media.github.cds.internal.unity3d.com/user/1745/files/009e9256-38ae-4bdb-bf60-e68ab7b79760)

After:
![shader-graph-get-effective-shader-stage-after](https://media.github.cds.internal.unity3d.com/user/1745/files/8e162c3c-c0f4-4fae-9073-1f0409350c35)

This is the total before/after. Before:
![shader-graph-before](https://media.github.cds.internal.unity3d.com/user/1745/files/c27f4877-2e26-49fe-85ce-44808ce2e794)

After:
![shader-graph-after](https://media.github.cds.internal.unity3d.com/user/1745/files/4b7b6065-db87-48af-b4b3-66bfb5a53fc8)