Optimizations to DecalSystem, DecalSet & DecalProjector
When profiling entering/exiting playmode using the Volvo Test Track project (https://github.com/Unity-Technologies/vwt-volvotesttrack/tree/main) it was observed that there was a large cost setting up decals.  This scene has 6,448 decal projectors in it so is a good stress test for the system - anecdotal evidence indicates this is not a unique situation with other Unity and user projects having been seen with similarly large numbers.

![image](https://media.github.cds.internal.unity3d.com/user/120/files/1573e283-e41c-4976-bc9d-f05f11eaa00f)

Digging deeper several optimization opportunities were identified to improve on the 526 ms initially being taken in the `DecalProjector::OnEnable()` function:

![image](https://media.github.cds.internal.unity3d.com/user/120/files/fa1e01e9-7ea0-46bb-ad1c-2b43f2385b82)

This PR implements the following improvements:

**1) Moved SetupMipStreamingSettings**

The function `SetupMipStreamingSettings()` was being called in `DecalSystem.AddDecal()` for every single decal to set up the mip streaming settings for the decal's material.  This is redundant much of the time however as decals tend to share materials - this PR includes an optimization to `DecalSystem.AddDecal()` so `SetupMipStreamingSettings()` is only be called when a material is seen for the first time rather than every time a decal is added.  

In the Volvo Test Track project this saved 10ms overall

**2) Changed Buffer Growth Strategy**

DecalSystem maintains multiple buffers of data for the decals it knows about, when `DecalSystem.AddDecal()` is called to add a new decal it has to grow these buffers if there isn't room to add any more.

The original strategy for this was to grow each buffer by `kDecalBlockSize` entries each time - a constant set to 128 so that's how many can then be added before further growth is required.  Each time this is done all existing entries have to be copied from the old full buffers to the newly created larger ones which of course takes time.

For the 6,448 decal projectors in the Volvo scene the buffers had to regrow 86 times each one being more expensive than the last.

This PR changes the growth strategy to instead grow proportionally by 20% each time clamping the growth factor to be not less than `kDecalBlockSize` (128) and not more than the new constant `kDecalMaxBlockSize` (2048).  This reduces the number of times growth is required and thus existing entries copied considerably, for this scene down from 86 to 20 reducing the enter/exit playmode time from 180ms to 65ms.  Most of this is saved in the expensive `DecalSet.ResizeJobArrays()` function.

**3) Changing how TransformAccessArray is Resized**

DecalSet originally was using `ResizeArray()` to increase the capacity of the `TransformAccessArray` when growing the job data buffers.  As `ResizeArray()` simply copies each transform in turn from the existing array to a new one internally it was calling the expensive `RemoveTransformFromGlobalTable()` and `AddTransformToGlobalTable()` functions even though the transform wasn't actually changing.  

This PR improves the performance by setting the capacity of the existing array directly rather than copying everything to a new one saving almost all the execution time.  This direct method can only grow the array not change it's size arbitrarily (the decal system only ever grows it here anyway) so the function `ResizeJobArrays()` was renamed to `GrowJobArrays()` to make this clearer.

Testing with the Volvo scene the cost when entering/exiting playmode reduced from 34ms to 0.5ms

**4) DecalProjector Event Handlers**

It was found that an excessive amount of memory allocation appeared to be happening in the `DecalProjector::OnEnable()` function.  This was found to be due to the registration of two separate event handlers for each decal projector listening for the `SceneVisibilityManager.visibilityChanged` and `PrefabStage.prefabStageOpened` events.  While overall memory use once in playmode was less than 5 MiB for these 6,448 projectors around 329 MiB was actually being allocated from over 58K individual allocations with the vast majority of it reverting to garbage.  This produced pressure on the managed heap resulting in garbage collection usually taking place in the function adding another ~259 ms to the cost.

![image](https://media.github.cds.internal.unity3d.com/user/120/files/01e9c2b8-b99e-4fa7-9c67-b6a9856c7cd1)

This PR changes the code so only one handler is registered for each event (when the first decal is enabled) and DecalProjector now maintains it's own list `m_DecalProjectorInstances` of active DecalProjector instances.  When the event handler is called it forwards this on to each instance in this list.  

By only having a single shared handler the number of allocations fell from 52K to 13.3K and the memory churn from 324 MiB to 3.2 MiB almost all of which is active.  Performance in the function also improved significantly due to far fewer heap calls reducing from 267ms (526ms including `GC.Collect`) to just 87ms

![image](https://media.github.cds.internal.unity3d.com/user/120/files/f73f5d20-691e-4942-849f-a7b2c2fec98d)