Remove URP RenderDataPatcher
Profiling with user projects showed that several hundred ms were being spent on domain reload in RendererDataPatcher whose sole function is to fix up UniversalRendererData Scriptable Objects that were made in Unity 2021.2 Alpha & Beta.  The cost of this code would scale with the size of the asset database - times up to nearly 500 ms were observed in one moderately sized user project.

This upgrader was created back when a mistake was made changing the GUID for the Universal Renderer, and it was to solve issues for people who had used the 21.2 alpha/beta.  As this was only alpha/beta builds of a version that is no longer supported anyway it's safe to remove this code and eliminate the domain reloading overhead.

(See **#devs-urp** thread here: https://unity.slack.com/archives/C89KFUUCT/p1678108080326519)

This PR simply removes this upgrader code saving the overhead