Mobile Optimizations for Static APV
Static APV was not initially designed to run on Mobile devices. This is a step to get it working within Mobile budgets. Much of the update revolves around:
- Using more `half` registers
- Reordering calculations to use more intrinsic functions (specifically `fma`, `dot`, etc.)
- Removing redundant casts between types (`float` -> `int` -> `float`, etc.)

There are still some things to look into (removing branches, investigating memory bandwidths for sampling the SH textures), but that is outside the scope of this initial PR.

This is a an update specifically for URP, however the modified code is in the Core SRP library, so HDRP will benefit from this as well.