[HDRP] Colored Translucent Material Type
Added a new material type **Colored Translucent** to support a RGB transmission mask.
This mode is specifically for thin double sided geometry.

To simplify things, this type doesn't require having a diffusion profile, so the index of refraction is set to 1, the thickness is unused (set to 0), world scale is set to 1, transmission mode is set to 'Thin Object' and transmission tint can be freely set from the pixel output in shadergraph. The color is stored on 20 bits so it has a good precision
NOTE: in the diffusion profile, transmission tint is HDR, but here it is clamped, artist feedback is that it is not an issue because this will usually be used for flowers or thin papers sheets that don't need to be HDR.

feedback doc from alix: https://docs.google.com/document/d/1p7gKpPJYq07G3Hm7rHvXkWrYQtHdlP6cX-zPlX0Y3_k/edit

![image](https://media.github.cds.internal.unity3d.com/user/2154/files/c547649b-1c48-424b-b641-1ef69067fe33)