Rpfoundation/materials leaking in corev2
Note To Reviewers: This is a recreated branch as the tooling was confused. Changes are identical to https://github.cds.internal.unity3d.com/unity/unity/pull/23992
Fix https://jira.unity3d.com/browse/UUM-27587
The bug reported several resources not being freed correctly by the core package during domain reloads. During adding of tests several other causes of not correctly freeing resources (unrelated to the core package) were discovered. I fixed a number of obvious ones in the URP and XR packages are part of this PR but one case remains in URP and one in the Font sytem. These are tracked as separate bugs:
https://jira.unity3d.com/browse/UUM-28555
https://jira.unity3d.com/browse/UUM-28697