[HDRP] Fix Color Pyramid History Buffer Logic
This change modifies the logic for the color pyramid history buffer
allocation to avoid attempting to allocate a history chain of zero
buffers. The underlying code does not expect or handle this behavior and
we should avoid it anyways to save memory.

The previous code would actually hit C# runtime errors in the rare case
where enough effects are disabled that a color pyramid is no longer
necessary.