[2023.1 backport] URP: Limit cmd.Blit usage to wireframe mode
Limit cmd.Blit usage to wireframe mode

cmd.Blit is only used in URP because of wireframe mode.
The issue is that without cmd.Blit, the core engine will replace the shader
used and draw in wireframe, making the full screen draw invisible.

This is incompatible with URP's CoreBlit because CoreBlit is made to be
used with the Blitter which uses geometryless Draw(3) calls with blit
scaling while cmd.Blit does Draw(4).

If Blitter is used in the scene view, it will work in shaded mode but not
in wireframe mode.
If cmd.Blit is used with a material, it will work in wireframe mode because
of shader replacement but it will not work in shaded mode because the
CoreBlit shader is not meant for Draw(4).
If cmd.Blit is used without a material, it will work in wireframe mode
but it will not use the uber post material in shaded mode among others.