[URP] Adding HDR output override per camera
HDR output currently depends on the render target format during post processing being HDR (so it doesn't clamp to values between 0 and 1), so it needs HDR rendering enabled. If HDR rendering is disabled on the render pipeline asset, HDR output is switched off and the display becomes SDR when possible. But this dependency wasn't intuitive when HDR rendering can be disabled per camera and it doesn't make sense to disable HDR output globally for one camera setup.

Instead this PR decouples HDR output and HDR rendering at the camera-level and introduces an HDR output override. Options are "Use Project Settings" (by default) and "Off". So individual camera can disable HDR rendering and maintain HDR output according to the project settings. Additionally a camera can purposely disable HDR output. Note this doesn't change the on-screen requirement for HDR output, offscreen cameras will not convert and/or encode the display color gamut.

<img width="538" alt="HDROutputWarningWithSDrRendering" src="https://media.github.cds.internal.unity3d.com/user/2846/files/581f4f02-2f1a-47c0-9b96-0e14d0607b36">

Additionally, cameras with HDR rendering disabled outputting to an HDR display will convert/encode the colors to the display's HDR color gamut now, fixing UUM-31589. It will attempt to output to HDR even if it renders incorrectly because of the post process render target format clamping the values. This only happens when color conversion is done in the Uber post pro pass, but color encoding has to be done in another pass (like Final post pro or Final Blit).

Here's the new behaviour matrix, still assuming an HDR display is available and active:
| HDR Rendering (URP Asset) | HDR Rendering (camera override) | HDR Output (camera override) | Actual Rendering | Actual Output (for the camera) | Actual Output (for the display)
| --- | --- | --- | --- | --- | --- |
| SDR | Use URP Asset | Use Project Settings | SDR Rendering | SDR Output | SDR Output |
| SDR | Use URP Asset | Off | SDR Rendering | SDR Output | SDR Output |
| SDR | Off | Use Project Settings | SDR Rendering | SDR Output | SDR Output |
| SDR | Off | Off | SDR Rendering | SDR Output | SDR Output |
| HDR | Use URP Asset | Use Project Settings | HDR Rendering | HDR Output | HDR Output |
| HDR | Use URP Asset | Off | HDR Rendering | SDR Output* | HDR Output |
| HDR | Off | Use Project Settings | SDR Rendering | HDR Output | HDR Output |
| HDR | Off | Off | SDR Rendering | SDR Output* | HDR Output |

In cases marked with *, the display is still in HDR mode but the URP doesn't do the necessary color conversion and encoding, so the image **will** be incorrect. But that's what the users should expect since they've force-disabled our HDR output processing on the camera.