[HDRP] Add dynamic resolution component
Users are asking for a way to control dynamic resolution. For now only settings are available but it doesn't help about how to write a proper manager for it. The example in our documentation is too restrictive. Here we provide a more complete manager as a script. It is self sufficient but users can still tweak it if needed.

Detail in documentation of dynamic resolution.