[2023.1] [URP] Fixed HDR output with final blit pass
Using the Universal Renderer with post processing disabled or the 2D renderer was broken with HDR output. It would not render, resulting in a black screen or staying on the last rendered frame. The bug was introduced by #21655, that changed what shader pass was being used in the Final Blit pass. But the shader pass index didn't exist for the blit material used when HDR output is active. This PR refactors the `BlitHDROverlay.shader` for debug draw support and modifies the Final Blit pass to use different shader pass indexes depending on the material used.

Additionally the HDR debug views were not getting the correct input, using the camera color attachment when they should be using the debug screen texture when HDR output is active. This was because "final" pass (Final Blit, Uber Post or Final Post) needed to call `ConfigureCameraColorTarget` to the debug screen texture.

On the render graph, the HDR debug views were culled because the final blit pass wasn't writing to the debug screen texture but directly to the backbuffer. The Final Post step was also causing a null reference exception, because it relied on the Uber Post step to be called first, which isn't always the case (e.g. base camera with PP enabled and upscaling + overlay camera with PP disabled).