2021.3/rpfoundation/materials leaking in corev2
2023.1 backport
This required some manual work and adding other URP changes that didn't seem to be backported.

--- original description
The bug reported several resources not being freed correctly by the core package during domain reloads. During adding of tests several other causes of not correctly freeing resources (unrelated to the core package) were discovered. I fixed a number of obvious ones in the URP and XR packages are part of this PR but one case remains in URP and one in the Font sytem. These are tracked as separate bugs:
https://jira.unity3d.com/browse/UUM-28555
https://jira.unity3d.com/browse/UUM-28697