[VFX] Add Template Window to create new Visual Effect assets
Jira: https://jira.unity3d.com/browse/VFXG-289
[Design Doc](https://docs.google.com/document/d/1RJn0zO2kstzbcJneL9Rf7DVJ7V9KvJywZrciTWm79KQ/edit?usp=sharing)
For doc review, here is the generated page: [VFX-Templates-doc.zip](https://github.cds.internal.unity3d.com/unity/unity/files/784/VFX-Templates-doc.zip)

⚠This is the second PR, the first one was a mess because of a bad commit which introduced 5000+ file changes.
Here is the link to the first PR: https://github.cds.internal.unity3d.com/unity/unity/pull/21087

The goal of this feature is to offer a relevant choice of templates when the user wants to create a new VFX asset.
This is achieved thanks to a new window that is displayed in every case:
- In the create menu in the project browser
- When clicking on the "Create new VFX asset" button in the "No asset" window
- When clicking on the "New" button in the inspector of a VFX component
- When inserting a template system in the current VFX graph window
<img width="800" alt="Unity_jOmJ0pHvlV" src="https://media.github.cds.internal.unity3d.com/user/4003/files/9116d906-339d-416b-a5ff-4084a3be64f2">


In addition to this new template window, a new "split button" has been added in the VFX Graph window toolbar.
This button gives a quick access to this new template window, either to create a completely new asset or insert a template in the currently edited VFX Graph.

<img width="297" alt="Unity_wEvptEMTcu" src="https://media.github.cds.internal.unity3d.com/user/4003/files/500bbc60-d349-40f2-90a6-c6060d4bac58">


It's also possible to create a custom template from any existing VFX asset.
But there's no fancy UI to do so on purpose because we need more input from GTF team before exposing any UI for this matter.
So today the easiest way is to:
- Select a VFX graph
- In the inspector there should be a section called `VFX Template`
- Fill in the fields
- Then, next time you open the template window, the custom template should appear in its dedicated category

<img width="443" alt="Unity_Bp7zgICmtP" src="https://media.github.cds.internal.unity3d.com/user/4003/files/f53a69de-2c85-4c02-a1d3-f0eaf6490243">