GFXRT-98 Decal support in HDRP Path Tracer
The current HDRP path tracer does not have for support decals, which are critical to match the visuals expected by users, as well as to match the capabilities of the competition. This PR adds support for Decals to the path tracer, by ensuring they are uploaded to the GPU when we are path tracing, and adding them to the ray tracing light cluster in that case.