[HDRP] Add Shape Column to Light Explorer
In HDRP, the Light Explorer doesn't let you change the light shape for Area and Spot lights. These are actually just different light types, but in the Inspector they are not listed as light types and can be changed in the Area and Spot light Shape settings. Additionally, URP Light Explorer has a Shape column for Area lights. For these reasons, a Shape column was added to the HDRP Light Explorer.

The new column in HDRP next to the Type column:
![image](https://media.github.cds.internal.unity3d.com/user/1926/files/2d317812-0405-404a-b769-5ffaeeadc75d)

How the same lights currently looks in URP (only support for Area Disc and Area Rectangle):
![image](https://media.github.cds.internal.unity3d.com/user/1926/files/d12fd6e0-1ff8-467f-9d37-bcc8438f8933)

https://jira.unity3d.com/browse/GFXRT-116