[RPW] Fix BitArray multiselection editing bug
Fix for https://jira.unity3d.com/browse/UUM-883. (more info available also at https://fogbugz.unity3d.com/f/cases/1411747)

The multi-editing bug was caused by maintaining separate SerializedProperty objects in a way that is usually not necessary, but in this case is required in order to display the "mixed" checkbox value.

With this PR, the SerializedBitArray logic has been unified to share a single implementation for all BitArray sizes. There is no longer a difference in behavior between single- and multi-editing.