[SRP][RPW]Stripping of the RenderPipelineGlobalSettings that do no belong to the pipeline that the build is generated for.
Currently, we are storing a map with RenderPipeline FullName and RenderPipelineGlobalSettings asset reference.

In a cross pipeline scenario, we will have multiple elements stored into the map, therefore we are sending resources from other pipelines.

We have implemented a new system that on `GraphicsSettings `we store only 1 reference to the asset that is filtered when building. In any other scenario we are accessing the map, but for standalone builds we are pointing directly to the asset that has been filtered before building the project.

The map, now is stored only on the `EditorOnlyGraphicsSettings`, making it more clean and more memory performant.