Disabled test of SpeedTree with lightmaps, they make little sense
The purpose of this PR is simply to disable testing which either makes no sense or which cannot be stable due to unstable input to the baking process. Follow up work will focus on making the input stable. Furthermore, when it's clear what kind of erronous data is fed into baking, it will be possible to improve input validation.