[2022.2 backport] [HDRP] Fix box-shape spotlight culling planes when used with BRG
Backport of #21150 for Entities Graphics.
JIRA issue: https://jira.unity3d.com/browse/GFXRT-121

This PR updates the culling planes for box-shaped spotlights to be consistent with their volume of influence. Currently box-shaped spotlights have culling planes that are still in a pyramid-shape, which can lead to missing shadow casters when used with BatchRendererGroup.