HDRP Colorchecker
The Color Checker is a tool used to calibrate lighting and post process.  It's an object that the user can add through GameObject/Rendering/Color Checker Tool. It is meant only as a production tool for lighting artists and won't be saved in build.

![image](https://media.github.cds.internal.unity3d.com/user/2204/files/1233f864-e528-443a-a3ed-94071fa6cd26)


It has several modes:

**Color Palette**
Default values are the data from Xrite color checker. 
![image](https://media.github.cds.internal.unity3d.com/user/2204/files/5bbd9e28-b7d4-48d7-b290-13f0c4c71b98)

The user can change those values as they please.
![image](https://media.github.cds.internal.unity3d.com/user/2204/files/fe35fa64-eda7-423c-9d4d-4390606acce9)

 It offers up to 64 values (internally, we are writing on a 8*8 texture)
![image](https://media.github.cds.internal.unity3d.com/user/2204/files/ac8ae2b3-89aa-4989-ad73-015bb0e45cc4)

**Cross Polarized Grayscale**
![image](https://media.github.cds.internal.unity3d.com/user/2204/files/9b37d067-dec9-448d-a0e6-7f6f044da66a)
It's the grayscale value from Xrite that have been probed with a cross polarized filter by @cyril  . It means it has been captured without specular reflections, making them more accurate in a PBR setting.

**Middle Gray**
Mid Gray value is used to calibrate exposure. 

**Reflection**
Only in spherical mode.
To quickly check on reflections.

**Stepped Luminance**
![image](https://media.github.cds.internal.unity3d.com/user/2204/files/9ec97b91-f809-49cb-88b6-aeefeba3eb66)
Useful to check on contrast and gamma. Pairs well with the gradient option.

**Material Palette**
![image](https://media.github.cds.internal.unity3d.com/user/2204/files/308d3e8c-7560-49a4-aeec-bf62ea958465)
Up to 12 materials that can be edited.
Default Values are (from bottom to top) : Snow, Charcoal, Iron, Copper, Silver, Gold

**External texture**
![image](https://media.github.cds.internal.unity3d.com/user/2204/files/cf395b91-e2ae-40ab-9731-61d123a1e5bc)
If procedural options fails, the user can feed their own external texture. 
There is the option to compare it with unlit values, especially useful if the user has access to an exr with Raw data from the reference photography.
The texture in the screenshot is from @rinaldo, and is not distributed, it is used as an example.

Gradient 
![image](https://media.github.cds.internal.unity3d.com/user/2204/files/3a0ff5b5-2b02-4262-8fc2-b055d0171bed)

Spherical Normals
![image](https://media.github.cds.internal.unity3d.com/user/2204/files/8cc89207-4edb-498b-b3ae-f06417db236d)

Compare to Unlit
![image](https://media.github.cds.internal.unity3d.com/user/2204/files/b01c9999-53dd-4a1e-98d7-4d3ef10c32c6)

Face to Camera
Makes the checker always face the view.
It is made on the shader so it's not supported by path tracer

## URP parity

The shader graph is using the Exposure Node that is not yet supported by URP. A pipeline check has been made to work around that, but will still produce warnings. Once lights units parity is achieved and this node supported, this tool will be able to be moved to Core and be available for both pipeline. 
If moved to Core, we could have still the issue of having a shader graph inside, to see depending on the discussions around cross pipeline shaders and where they should be.