[RPW] Fix volume confirmation box when opening project settings
This PR fixes this checkbox that could appear under certain circumstances:

![image](https://media.github.cds.internal.unity3d.com/user/3380/files/ab2df1db-1033-46d9-ad78-bae4b2990cb0)

Specifically, this could appear if you have deleted the default volume profile from disk and it must be recreated. In this case, we don't ask user for confirmation because there is no risk that we are messing up a profile they are using elsewhere.