[URP] Changed ScriptableRenderPass.Execute from abstract to virtual
Jira Epic:
https://jira.unity3d.com/browse/URP-1297

As part of the RenderGraph work, since now ScriptableRenderPass can support different render path, there is not anymore the hard requirement for every pass to override the Execute() base function.
This PR loosens that requirement improving as a QOL improvement for users that are interested in writing passes only for the RenderGraph path.
In case the non-RG pass is executed, any pass without an Execute override will just throw a warning, for consistency with the RG path behaviour.

This API change is non disruptive and safe and already discussed and approved by the scripting team in this slack thread:
https://unity.slack.com/archives/C06TPSM32/p1682414940923909