[URP] Fix HDR debug view edge-case with Render Graph
This PR fixes an edge-case, with HDR output active or not, when:
- RenderGraph is enabled
- post-processing is enabled
- there's no antialiasing (no final post processing)
- the user has injected passes in `RenderPassEvent.AfterRenderingPostProcessing`
- HDR debug views are active (HDR Debug Mode in the Rendering Debugger)

The reason was the uber post processing rendering to backbuffer when it should have been rendering to the debug screen texture meant to override it. The render graph logic to override the target of Uber post processing was checking the wrong parameters.

Additionally this PR fixes the pass injection in AfterRenderingPostProcessing, it was being executed after Final Post Processing when using the RenderGraph, rather than before like the documentation and the non-Render Graph path suggested.