[VFX] Fix GPU hang caused by VFX shadows being rendered before the VFX PrepareCamera commands are executed
The order of execution between shadow map rendering and VFX prepare camera commands was incorrect in this version.
After backporting [a fix PR](https://github.cds.internal.unity3d.com/unity/unity/pull/20023), there was an unexpected difference in the behavior of 2022.2 compared to trunk, with those two parts using different command buffers.
The suggested fix is to force the execution of the prepare camera commands command buffer, only in this version, because using the same command buffer as it is done in trunk could have other side effects.