URP Forward+ Reflection Probe Node flickering
When using the Reflection Probe Node in Forward+, the object will be flickering heavily. This happens because the underlying code for that node doesn't go via the regular reflection probe path. Instead, it just samples the first bound reflection probe. Since this is never bound in Forward+, it seems to end up accessing garbage.

To fix this, we now only sample the environment map in Forward+ for that node. We cannot sample local probes, because there is no position input to the node. But at least this will stop the flickering, and give a somewhat reasonable return value.

Note that in Forward, this node doesn't support probe blending or box projection either. Both features would need a position input. That would also allow it to sample local reflection probes in Forward+. I think that should be done later on though, as we need to modify the node for that to work, leading to some upgrade concerns.