[URP] Sorting of screen space decals fix
Screen space decals used a camera position dependent sorting, which caused bugs when moving the camera around overlapping decals sharing the same material.

This PR changes the sorting criteria to the same used by DBuffer decals. A previous fix was pushed for this issue but turned out to not fix it completely in all cases, so this changes the sorting criteria again also making sure to match the same sorting used by the DBuffer technique for a more consistent look across different implementations.